Relocate Postgres Casting.
